+++ backend/product/src/main/java/org/samtuap/inong/domain/farm/api/FarmController.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.samtuap.inong.domain.farm.dto.*;
 import org.samtuap.inong.domain.farm.service.FarmService;
 import org.samtuap.inong.domain.seller.dto.FarmCategoryResponse;
+import org.samtuap.inong.domain.seller.dto.SellerFarmInfoUpdateRequest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -88,6 +89,12 @@ public class FarmController {
                                          @RequestHeader("sellerId") Long sellerId) {
         FarmCreateResponse response = farmService.createFarm(request, sellerId);
         return new ResponseEntity<>(response, HttpStatus.CREATED);
+    }
+
+    @PostMapping("/update")
+    public ResponseEntity<?> updateFarmInfo(@RequestHeader("sellerId") Long sellerId, @RequestBody SellerFarmInfoUpdateRequest infoUpdateRequest){
+        farmService.updateFarmInfo(sellerId, infoUpdateRequest);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("/categories")
